requirements.txt
streamlit
openai
azure-cognitiveservices-speech
python-dotenv
requests